Split up navigation index
