+++ apps/arcade-city/src/navigation/LinkingConfiguration.ts
@@ -4,10 +4,9 @@
  * https://reactnavigation.org/docs/configuring-links
  */
 
-import { LinkingOptions } from '@react-navigation/native'
 import * as Linking from 'expo-linking'
-
-import { RootStackParamList } from '../types'
+import { LinkingOptions } from '@react-navigation/native'
+import { RootStackParamList } from './types'
 
 const linking: LinkingOptions<RootStackParamList> = {
   prefixes: [Linking.createURL('/')],
