+++ apps/arcade-city/src/navigation/index.tsx
@@ -1,0 +1,17 @@
+/**
+ * If you are not familiar with React Navigation, refer to the "Fundamentals" guide:
+ * https://reactnavigation.org/docs/getting-started
+ */
+import * as React from 'react'
+import { navTheme } from 'views/theme'
+import { NavigationContainer } from '@react-navigation/native'
+import LinkingConfiguration from './LinkingConfiguration'
+import { RootNavigator } from './root-navigator'
+
+export default function Navigation() {
+  return (
+    <NavigationContainer linking={LinkingConfiguration} theme={navTheme}>
+      <RootNavigator />
+    </NavigationContainer>
+  )
+}
