+++ apps/arcade-city/src/navigation/root-navigator.tsx
@@ -1,0 +1,33 @@
+import { color, typography } from '@arcadecity/ui'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
+import { RootStackParamList } from './types'
+
+/**
+ * A root stack navigator is often used for displaying modals on top of all other content.
+ * https://reactnavigation.org/docs/modal
+ */
+const Stack = createNativeStackNavigator<RootStackParamList>()
+
+export function RootNavigator() {
+  return (
+    <Stack.Navigator>
+      <Stack.Screen name='Root' component={BottomTabNavigator} options={{ headerShown: false }} />
+      <Stack.Screen name='NotFound' component={NotFoundScreen} options={{ title: 'Oops!' }} />
+      <Stack.Group screenOptions={{ presentation: 'modal' }}>
+        <Stack.Screen
+          name='Modal'
+          component={ModalScreen}
+          options={{
+            headerStyle: {
+              backgroundColor: color.tabbar,
+            },
+            headerTitleStyle: {
+              color: color.text,
+              fontFamily: typography.secondary,
+            },
+          }}
+        />
+      </Stack.Group>
+    </Stack.Navigator>
+  )
+}
