+++ apps/arcade-city/src/navigation/tab-navigator.tsx
@@ -1,69 +1,8 @@
-/**
- * If you are not familiar with React Navigation, refer to the "Fundamentals" guide:
- * https://reactnavigation.org/docs/getting-started
- */
-import * as React from 'react'
-import { ColorSchemeName, Pressable, View, ViewStyle } from 'react-native'
+import { Pressable, View, ViewStyle } from 'react-native'
 import { color, palette, typography } from '@arcadecity/ui'
 import { FontAwesome, FontAwesome5 } from '@expo/vector-icons'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
-import { DarkTheme, NavigationContainer } from '@react-navigation/native'
-import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import Colors from '../constants/Colors'
-import ModalScreen from '../screens/ModalScreen'
-import NotFoundScreen from '../screens/NotFoundScreen'
-import TabOneScreen from '../screens/TabOneScreen'
-import TabTwoScreen from '../screens/TabTwoScreen'
-import {
-  RootStackParamList, RootTabParamList, RootTabScreenProps
-} from '../types'
-import LinkingConfiguration from './LinkingConfiguration'
-
-const navTheme = {
-  colors: {
-    ...DarkTheme.colors,
-    background: color.background,
-  },
-  dark: true,
-}
-
-export default function Navigation() {
-  return (
-    <NavigationContainer linking={LinkingConfiguration} theme={navTheme}>
-      <RootNavigator />
-    </NavigationContainer>
-  )
-}
-
-/**
- * A root stack navigator is often used for displaying modals on top of all other content.
- * https://reactnavigation.org/docs/modal
- */
-const Stack = createNativeStackNavigator<RootStackParamList>()
-
-function RootNavigator() {
-  return (
-    <Stack.Navigator>
-      <Stack.Screen name='Root' component={BottomTabNavigator} options={{ headerShown: false }} />
-      <Stack.Screen name='NotFound' component={NotFoundScreen} options={{ title: 'Oops!' }} />
-      <Stack.Group screenOptions={{ presentation: 'modal' }}>
-        <Stack.Screen
-          name='Modal'
-          component={ModalScreen}
-          options={{
-            headerStyle: {
-              backgroundColor: color.tabbar,
-            },
-            headerTitleStyle: {
-              color: color.text,
-              fontFamily: typography.secondary,
-            },
-          }}
-        />
-      </Stack.Group>
-    </Stack.Navigator>
-  )
-}
+import { RootTabParamList, RootTabScreenProps } from './types'
 
 /**
  * A bottom tab navigator displays tab buttons on the bottom of the display to switch screens.
@@ -71,7 +10,7 @@ function RootNavigator() {
  */
 const BottomTab = createBottomTabNavigator<RootTabParamList>()
 
-function BottomTabNavigator() {
+export function BottomTabNavigator() {
   return (
     <BottomTab.Navigator
       initialRouteName='TabOne'
