+++ apps/arcade-city/src/navigation/types.ts
@@ -4,7 +4,9 @@
  */
 
 import { BottomTabScreenProps } from '@react-navigation/bottom-tabs'
-import { CompositeScreenProps, NavigatorScreenParams } from '@react-navigation/native'
+import {
+  CompositeScreenProps, NavigatorScreenParams
+} from '@react-navigation/native'
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
 
 declare global {
