+++ apps/arcade-city/src/views/theme/index.ts
@@ -1,1 +1,2 @@
 export * from './images'
+export * from './nav'
